refactor: create presentation, domain and data packages
