+++ app/src/main/java/io/thomasgasangwa/bookstore/presentation/MainActivity.kt
@@ -1,5 +1,6 @@
-package io.thomasgasangwa.bookstore
+package io.thomasgasangwa.bookstore.presentation
 
+import BookStoreTheme
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
@@ -11,7 +12,6 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import io.thomasgasangwa.bookstore.ui.theme.BookStoreTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
