+++ app/src/main/java/io/thomasgasangwa/bookstore/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package io.thomasgasangwa.bookstore.ui.theme
+package io.thomasgasangwa.bookstore.presentation.ui.theme
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.Font
